+++ datasophon-common/src/main/java/com/datasophon/common/dto/FrameServiceRoleDTO.java
@@ -33,7 +33,7 @@ import java.util.List;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record FrameServiceRoleDTO(
         Long id,
-        Integer serviceId,
+        Long serviceId,
         String serviceName,     // 新增：服务名称字段
         String serviceRoleName,
         Integer serviceRoleType,
@@ -52,7 +52,7 @@ public record FrameServiceRoleDTO(
     /**
      * 创建基础FrameServiceRoleDTO，不包含主机列表
      */
-    public static FrameServiceRoleDTO of(Long id, Integer serviceId, String serviceName, String serviceRoleName,
+    public static FrameServiceRoleDTO of(Long id, Long serviceId, String serviceName, String serviceRoleName,
             Integer serviceRoleType, String cardinality, String frameCode) {
         return new FrameServiceRoleDTO(id, serviceId, serviceName, serviceRoleName, serviceRoleType, cardinality,
                 null, null, frameCode, null, null, null);
@@ -61,7 +61,7 @@ public record FrameServiceRoleDTO(
     /**
      * 创建包含主机列表的FrameServiceRoleDTO
      */
-    public static FrameServiceRoleDTO withHosts(Long id, Integer serviceId, String serviceName, String serviceRoleName,
+    public static FrameServiceRoleDTO withHosts(Long id, Long serviceId, String serviceName, String serviceRoleName,
             Integer serviceRoleType, String cardinality, String serviceRoleJson, String serviceRoleJsonMd5,
             String frameCode, String jmxPort, String logFile, List<String> hosts) {
         return new FrameServiceRoleDTO(id, serviceId, serviceName, serviceRoleName, serviceRoleType, cardinality,
